+++ ntnx_api/client.py
@@ -8,8 +8,13 @@ ntnx_api.client
 
 ApiClient
 ^^^^^^^^^
+.. autoclass:: ntnx_api.client.PrismApi
+    :members:
+    :undoc-members:
+
 .. autoclass:: ntnx_api.client.ApiClient
     :members:
+    :undoc-members:
 
 .. autoclass:: ntnx_api.client.NutanixError
 
@@ -134,8 +139,7 @@ class NutanixRestHTTPError(NutanixError):
 @deprecated(
     reason="""This ApiClient class is being deprecated in favor of separate classes for teach Nutanix endpoint type being connected with. This will simplify
     future development and use and allow for more granular changes based on the requirements of the endpoints API.
-
-       - Prism API connection management has moved to client.Prism
+    Prism Element or Central API connection management has moved to :class:`.PrismApi`
     """,
     version='1.1.0',
 )
@@ -350,8 +354,8 @@ class ApiClient(object):
 
 
 @versionadded(
-    reason="""
-    This class superseeds client.ApiClient. To interact with the Prism Element or Prism Central APIs please use this class.
+    reason="""This class supersedes the originally released class :class:`.ApiClient`.
+    To interact with the Prism Element or Prism Central APIs please use this class instead of the deprecated class.
     """,
     version='1.1.0',
 )